update: Refactoring.
